Reduce confusion at intro of 00_architecture_overview.md
I found the bullet points to be confusing because it wasn't clear these are the same things you've just pointed to in the table. I also found myself frustrated that I didn't know what is an algorithm, module, etc in this context so staring at the table wasn't helping me (since one cannot see the later sections that will soon explain).